krb5: only try pkinit with Smartcard credentials
Currently pkinit is tried if a Smartcard is present. But depending on
the used PAM service and other configurations it might happen that the
user didn't provide the Smartcard PIN but e.g. the password. Hence,
before trying pkinit we should check if the right credentials are
available.

Resolves:
https://github.com/SSSD/sssd/issues/5290

Reviewed-by: Alexey Tikhonov <atikhono@redhat.com>
(cherry picked from commit bca413267f58395e22415edc662a7ba89fbe7b30)