ad: add required 'cn' attribute to subdomain object
If the forest root is not part of the return trusted domain objects
from the local domain controller we generate an object for further
processing. During this processing it is expected that the 'cn'
attribute is set and contains the name of the forest root. So far this
attribute was missing and it is now added by this patch.

Resolves: https://github.com/SSSD/sssd/issues/5926

Reviewed-by: Pavel Březina <pbrezina@redhat.com>
(cherry picked from commit bf6059eb55c8caa3111ef718db1676c96a67c084)