SDAP: Include struct ldb_dn in struct sdap_search_base
struct sdap_search_base includes the DN as a string. To better
compare the DNs it is better to use a struct ldb_dn, in addition to
the string.

The struct ldb_dn also needs to keep the associated struct ldb_context,
so we are also storing it in the structure.

Resolves: https://github.com/SSSD/sssd/issues/5708

Reviewed-by: Sumit Bose <sbose@redhat.com>
Reviewed-by: Tomáš Halman <thalman@redhat.com>
(cherry picked from commit cfc591d6519e73572eb7750770b5e6ce3ba9a1d4)

Reviewed-by: Alexey Tikhonov <atikhono@redhat.com>